split out utils from types file
